+++ changelog.md
@@ -1,5 +1,6 @@
 ### Version 0.1.2
 - Setting the size of the component to match the size of selected background
+- Update style of inputs with nested buttons and inputs
 ### Version 0.1.1
 - Custom Animation can be added
 - Adjustable width of Animated Banner's parent